+++ src/core/config/config.interfaces.ts
@@ -1,5 +1,7 @@
 import type { Buffer } from 'node:buffer'
 
+import { MatterConfig } from '../matter/matter.interfaces.js'
+
 export interface StartupConfig {
   host?: '::' | '0.0.0.0' | string
   httpsOptions?: {
@@ -24,9 +26,7 @@ interface PluginChildBridge {
     DEBUG?: string
     NODE_OPTIONS?: string
   }
-  matter?: {
-    port?: number
-  }
+  matter?: MatterConfig
 }
 
 export interface PlatformConfig {
@@ -54,9 +54,7 @@ export interface HomebridgeConfig {
     manufacturer?: string
     model?: string
     firmwareRevision?: string
-    matter?: {
-      port?: number
-    }
+    matter?: MatterConfig
   }
   mdns?: {
     interface?: string | string[]
@@ -70,13 +68,6 @@ export interface HomebridgeConfig {
   accessories?: AccessoryConfig[]
   plugins?: string[]
   disabledPlugins?: string[]
-}
-
-/**
- * Matter configuration interface
- */
-export interface MatterConfig {
-  port?: number
 }
 
 /**